Correct the handling of empty labels.
Correct the handling of empty labels.  PythonUI erroneously closed boxes even if
they were opened without a label, in which case closeBox() could be called on
the last box.  Now PythonUI keeps track of boxes opened with an empty label and
does not pop the current name-space then.  This removes the need to check
whether the last box is being closed.