Fix cffi caching for DSP files.
Fix cffi caching for DSP files, but not for inline DSP code strings.  The
problem was that the hash of the preamble passed to cffi.verify() would change
because the C file that FAUST outputs to has a random name (well, duh, I wrote
it that way).  So instead of #include-ing the C file in the preamble, simply
concatenate the file contents to it.

This doesn't fix the issue for inline DSP code strings because the first call to
openBox() passes the file name as an argument.