FAUSTPy: migrate to new FAUST llvm-c-dsp C API.
Migrate to the new FAUST llvm-c-dsp C API.  "llvm-dsp" is part of libfaust and
is basically an LLVM-based embeddable FAUST compiler that returns C++ objects
representing the compiled DSPs.  "llvm-c-dsp" is a C API built on top of that.
Apart from being a nicer approach than compile_faust(), this also makes it
potentially possible to create binary distributions of FAUSTPy.

Also update the test suite to the new FAUSTPy classes.